Model loading is eager in production, lazy in development
