+++ Dockerfile
@@ -34,7 +34,6 @@ COPY pyproject.toml .
 COPY supervisord.conf /etc/supervisord.conf
 
 RUN pip install --default-timeout=120 -e .
-RUN mypy
 
 ENV FLASK_APP=src.seer.app
