+++ src/seer/bootup.py
@@ -25,6 +25,7 @@ def bootup(
     init_migrations=False,
     init_db=True,
     with_async=False,
+    eager_load_inference_models=False,
 ) -> Flask:
     sentry_sdk.init(
         dsn=os.environ.get("SENTRY_DSN"),
@@ -35,6 +36,14 @@ def bootup(
     )
     app = Flask(name)
     app.config["SQLALCHEMY_DATABASE_URI"] = os.environ["DATABASE_URL"]
+
+    from seer.inference_models import cached
+
+    if eager_load_inference_models:
+        for item in cached:
+            # Preload model
+            item()
+
     if init_db:
         db.init_app(app)
         if init_migrations:
@@ -43,4 +52,5 @@ def bootup(
             Session.configure(bind=db.engine)
             if with_async:
                 AsyncSession.configure(bind=create_async_engine(db.engine.url))
+
     return app
